+++ addOns/authhelper/CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 - Add configuration support for the wait time after Client Script Based Authentication.
 - Include the Web Element being interacted with in the Client Script Based Authentication diagnostics.
+- Allow to enable authentication diagnostics for Client Script and Browser Based Authentication through the GUI.
 
 ### Changed
 - Warn when the recorded script used with Client Script Based Authentication does not launch a browser.